refactor: Split kanban constants and update team labels
